Update JHBridge and JHArduino library to remove variable names
Update JHBridge and JHArduino library to remove variable names and use
an enum instead. All JeeHell parameters are in parameters.json. This
file is the reference, from it, the parameters.h of JHArudino is
generated. parameter.json is also used by JHBridge to generate JeeHell
FMGSserver.iocp.ini file for a given board configuration (boards must be
connected at the generation time to know which parameter are used and
must be provided by JH SIOC interface)